requirements.txt
asgiref==3.9.1
Django==5.2.6
django-filter==25.1
django-rest-framework==0.1.0
djangorestframework==3.16.1
drf-yasg==1.21.10
inflection==0.5.1
packaging==25.0
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
uritemplate==4.2.0
